# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.MobileMoneyRequest:
    properties:
      amount:
        example: 50000
        type: number
      description:
        example: Property payment
        type: string
      phone_number:
        example: "+250788123456"
        type: string
      provider:
        enum:
        - mtn
        - airtel
        example: mtn
        type: string
    required:
    - amount
    - phone_number
    - provider
    type: object
  handlers.PropertySearchRequest:
    properties:
      district:
        example: Kigali
        type: string
      latitude:
        example: -1.9536
        type: number
      longitude:
        example: 29.8739
        type: number
      max_price:
        example: 10000000
        type: number
      max_size:
        example: 1000
        type: number
      min_price:
        example: 1000000
        type: number
      min_size:
        example: 100
        type: number
      page:
        example: 1
        type: integer
      page_size:
        example: 50
        type: integer
      property_type:
        example: residential
        type: string
      radius:
        example: 5000
        type: number
      sort_by:
        example: created_at
        type: string
      sort_order:
        example: desc
        type: string
    required:
    - latitude
    - longitude
    type: object
...
  utils.APIError:
    properties:
      details:
        type: string
      message:
        type: string
    type: object
  utils.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/utils.APIError'
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: api.landvaluationsystem.rw
info:
  contact:
    email: support@landvaluationsystem.rw
    name: API Support
    url: http://www.landvaluationsystem.rw/support
  description: Rwanda Land Valuation System with Official Gazette Integration
  license:
    name: Proprietary
    url: http://www.landvaluationsystem.rw/license
  termsOfService: http://swagger.io/terms/
  title: Land Valuation System API
  version: "1.0"
paths:
...
  /api/v1/properties/search:
    post:
      consumes:
      - application/json
      description: Search for properties near a location with optional filters for
        price, type, and district
      parameters:
      - description: Search filters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PropertySearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Properties found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Search properties by location and filters
      tags:
      - properties
...
  /payments/{transaction_id}/status:
    get:
      consumes:
      - application/json
      description: Retrieve the current status of a payment transaction
      parameters:
      - description: Transaction ID
        in: path
        name: transaction_id
        required: true
        type: string
      - description: Payment provider (mtn|airtel)
        in: query
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment status retrieved
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Transaction not found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Check payment status
      tags:
      - payments
  /payments/callback:
    post:
      consumes:
      - application/json
      description: Process payment provider callbacks (webhook)
      parameters:
      - description: Payment provider (mtn|airtel)
        in: query
        name: provider
        required: true
        type: string
      - description: Callback payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Callback processed
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Invalid callback
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Handle payment callback
      tags:
      - payments
  /payments/mobile-money:
    post:
      consumes:
      - application/json
      description: Initiate a payment via MTN or Airtel mobile money with validation
      parameters:
      - description: Payment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.MobileMoneyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Payment initiated
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Invalid request or validation failed
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Initiate mobile money payment
      tags:
      - payments
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"